Refactor v2 events as its own module.
