+++ src/deploy/functions/services/index.ts
@@ -1,5 +1,6 @@
 import * as backend from "../backend";
 import * as iam from "../../../gcp/iam";
+import * as v2events from "../../../functions/events/v2";
 import { obtainStorageBindings, ensureStorageTriggerRegion } from "./storage";
 
 const noop = (): Promise<void> => Promise.resolve();
@@ -37,7 +38,7 @@ export const StorageService = {
 };
 
 /** Mapping from event type string to service object */
-export const EVENT_SERVICE_MAPPING: Record<string, any> = {
+export const EVENT_SERVICE_MAPPING: Record<v2events.Event, Service> = {
   "google.cloud.pubsub.topic.v1.messagePublished": PubSubService,
   "google.cloud.storage.object.v1.finalized": StorageService,
   "google.cloud.storage.object.v1.archived": StorageService,
@@ -55,5 +56,5 @@ export function serviceForEndpoint(endpoint: backend.Endpoint): Service {
     return NoOpService;
   }
 
-  return EVENT_SERVICE_MAPPING[endpoint.eventTrigger.eventType] || NoOpService;
+  return EVENT_SERVICE_MAPPING[endpoint.eventTrigger.eventType as v2events.Event] || NoOpService;
 }
